+++ src/models/types/TokensAndBombs.ts
@@ -1,0 +1,6 @@
+import type { Raid } from "./Raid";
+
+export interface TokensAndBombs {
+    bombs: Raid[];
+    tokens: Raid[];
+}